backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn

# Google Cloud Client Libraries
google-cloud-firestore
google-cloud-speech
google-cloud-texttospeech
google-auth
google-auth-oauthlib
google-api-python-client

# Environment Variables
python-dotenv
# HTTP client for async calls (optional)
httpx

# Optional: Pydantic for model validation (already included via FastAPI)
pydantic